+++ tools/test.py
@@ -30,7 +30,8 @@ if __name__ == '__main__':
     for attribute in default_values.keys():
         if config[attribute] is None:
             config[attribute] = default_values[attribute]
-    config['test'] = True
+    if not config['inference'] and not config['test']:
+        config['test'] = True
 
     task = Task.init(project_name='simvp/test', task_name=config['ex_name']+"_test")
     task.connect_configuration(config)
@@ -41,6 +42,9 @@ if __name__ == '__main__':
     exp = BaseExperiment(args, task)
     rank, _ = get_dist_info()
 
-    mse = exp.test()
-    if rank == 0 and has_nni:
+    if config['inference'] and not config['test']:
+        mse = exp.inference()
+    else:
+        mse = exp.test()
+    if rank == 0 and has_nni and mse is not None:
         nni.report_final_result(mse)